+++ app/javascript/users/pages/course_events_index.vue
@@ -1,8 +1,11 @@
 <template>
   <v-container fill-height class="d-flex justify-center">
 
-    <!--CourseEventsListCards v-if="subscribed_course_events.length > 0" :course_events="subscribed_course_events" /-->
-    <CourseEventsCalendar v-if="subscribed_course_events.length > 0" :course_events="subscribed_course_events" />
+    <CourseEventsTable :course_events="course_events" @select-course_event="setCourseEvent"/>
+    <h3 v-if="course_events.length == 0 && !loading"
+        class="d-flex justify-center subtitle">{{ $t('course_event.list.none') }}</h3>
+
+    <CourseEventsListCards v-if="subscribed_course_events.length > 0" :course_events="subscribed_course_events" />
     <h3 v-else-if="subscribed_course_events.length == 0 && !loading"
         class="d-flex justify-center subtitle">{{ $t('course_event.list.none_subscribed') }}</h3>
     <v-card v-else
@@ -19,8 +22,8 @@
 
 <script>
   // Components
-  // import CourseEventsListCards from '../components/course_events/course_events_list_cards'
-  import CourseEventsCalendar from '../components/course_events/course_events_calendar'
+  import CourseEventsListCards from '../components/course_events/course_events_list_cards'
+  import CourseEventsTable from '../components/course_events/course_events_table'
 
   import { mapState, mapActions } from 'vuex'
 
@@ -29,8 +32,8 @@
       this.fetchCourseEvents()
     },
     components: {
-      // CourseEventsListCards,
-      CourseEventsCalendar,
+      CourseEventsListCards,
+      CourseEventsTable,
     },
     data() {
       return {