+++ app/javascript/users/router.js
@@ -12,6 +12,7 @@ import EditProfile from './components/profile/edit'
 import NotFound from './pages/not_found'
 import Home from './pages/home'
 import Dashboard from './pages/dashboard'
+import CourseEventIndex from './pages/course_events_index'
 import CourseEventShow from './pages/course_event_show'
 import ProductsIndex from './pages/products_index'
 import ProductShow from './pages/product_show'
@@ -31,6 +32,11 @@ const routes = [
   },
   {
     path: '/dashboard',
+    name: 'dashboard',
+    component: Dashboard,
+  },
+  {
+    path: '/courses',
     name: 'dashboard',
     component: Dashboard,
   },
@@ -99,12 +105,12 @@ const routes = [
     component: EditProfile
   },
   {
-    path: '/course_event',
-    name: 'editProfile',
-    component: EditProfile
+    path: '/courses',
+    name: 'CourseEventIndex',
+    component: CourseEventIndex
   },
   {
-    path: '/course_event/:id',
+    path: '/course/:id',
     name: 'courseEventShow',
     component: CourseEventShow
   },